HUE-219. Make Hue work with CDH3 trunk
Many changes to make Hue work against the latest build of CDH3. The most
substantial changes are to the plugins and to the Jobsub interposition code.
Note that this commit does not address any Beeswax changes, and in fact removes
Beeswax from the apps/Makefile. That work will come in a subsequent commit.