HUE-417. Beeswax views with big white bar at the bottom
This was a change with how SplitView calculates its target size. Previously we always filled the window, but to allow SplitView to nest for multi-split layouts we had to switch to measuring the parent element of the splitview element. This introduced this sizing problem. The new requirement is that the splitview has to be the root element or, if not, the parent of the splitview needs to have a SizeTo filter to configure its height. This happens in our code often when we have a div.view in the root. The simple fix, when possible, is to also make this div the splitview element. I've adjusted this in all our current usages.