HUE-550. Switching to LIFO queue for Thrift connection pool.
This was a subtle bug and a fun debugging experience.  It's not every
day you get to use YourKit, wireshark, strace, and log-based tools at
the same time!

I noticed, while doing some other testing, that the filebrowser was very
slow, but only sometimes.  In debugging, I found that every second
refresh, roughly, was slow.  Though I had originally attributed this
slowness to the UI, the logs clearly showed that some Thrift calls were
taking almost 5 seconds.

Hue has a connection pool for Thrift clients.  This pool had a default
size of 10.  Meanwhile, the default pool size on the namenode Thrift
plugin was 5.  So, what would happen is that a Thrift client would be
requested, and the pool would round robin to the next client.  Client c1
would lock up t1 on the server.  c2 would lock up t2.  And so on till c5
and t5.  Then, when c6 came around, it would wait until the server had a
spare thread.  This would free up as soon as c1 would time out (5
seconds!).

The solution is two-fold: (1) instead of FIFO for the client pool,
we switch to a LIFO solution.  Then the same socket is constantly
re-used.  (2) We up the thread pool size to be the same as the
connection pool size by default.

There were several moderately interesting things this analysis
ruled out:
 (1) The namenode is not slow. We do a handful of
 operations for every file browser page, which isn't
 great, but it's not slow.  YourKit didn't even
 sample a single stack trace; nothing was going on.
 (2) We thought there was a problem in the reconnection
 logic.  It turns out (as the comment points out) that
 select() will return a socket as readable if it's
 been closed or timed out.  This is all and good; that
 code seems to be working correctly.
 (3) Funky TCP stuff. I was worried that somehow
 connection re-use was broken. It seems fine; wireshark
 worked, but didn't say anything too interesting.

In terms of debugging, the major point was isolating
the call by writing the following script.  If I'd
done this an hour earlier (instead of hitting refresh
on a browser), this would have taken less time.
Kudos to Todd for mkaing me do it.

  import time
  import os

  os.environ["DJANGO_SETTINGS_MODULE"] = 'desktop.settings'

  from desktop.lib import fsmanager
  fs = fsmanager.get_default_hdfs()[1]
  fs.setuser("hdfs")

  i = 0
  while True:
    i += 1
    now = time.time()
    fs.stats("/")
    diff = time.time() - now
    print i, diff

This clearly showed a pattern of 5 instantaneous calls,
followed by one that took 5 seconds, followed by another 4
quick calls, etc.  After that, I recognized that adjusting
the server pool size to 50 eliminated the problem, and
adjusting it to 1 exacerbated it.

Kudos to Todd for the insight that the queue was working
backwards, and to Henry for his assistance as well.
(cherry picked from commit 51630f42982c5d5b887f09c76f52b70fcd78d0d8)