Edit groups by permission
This adds a tab in the User Administrator to allow for assigning permissions by
selecting a set of groups that have a particular permission. Depending on the
use case, it makes sense to be able to edit permissions by modifying them for a
particular group, as well as looking at a particular permission, and the groups
that have that permission.