Adding models to support LDAP integration
This adds an LdapGroup model, which will be used to denote groups that have
been imported from an LDAP system, as well as adds a creation_method field to
the UserProfile model, which will identify a user as a Hue-created user or a
user imported from an external system. Additionally, the authentication
backends which import users from external systems have been augmented to set
the field.