Add a configurable default user group
This adds a configuration parameter to the useradmin application, which allows
a default group to be specified. The group will be created automatically, and
may not be deleted (although if the name of the configured group is changed,
a previous default group may be deleted). When adding a user, or logging in via
the LDAP or PAM authentication backends, this group will be selected as a
default group for the user.