[hadoop] Add config for Yarn cluster, Mapred port, and `submit_to' param
The `yarn_clusters' is a new section. The plural is to be consistent with HDFS
and MR config. But there is no plan to work with multiple Yarn clusters.

The JT port needs to be configured, in order for Oozie to know how to connect
to it.

Added a new `submit_to' boolean param, to mark which cluster to run jobs on.