HUE-788 [core] Enable apps to be url namespaced
When we register the URLs of an app with a namespace,
all the reversed views will need to have the namespace of the app.

If the namespace in the url is not specified
(e.g. 'list_dir' instead of 'filebrowser:list_dir') the 'list_dir'
view does not exist for Django as it was registered from urls.py
with the namespace 'filebrowser'. In consequence all the urls of
an app need to be converted with the namespace before switching
'IS_URL_NAMESPACED' to True (or a "Reverse for 'view' with arguments '()'"
will show up.).