HUE-797 [useradmin] Admin cannot delete a user account.
The list_users.mako template had a javascript function that created a popup.
The template for the popup was taken and put into delete_user.mako.
Then, the delete_user view was updated to include appropriate context vars.
New javascript code was added to appropriately call an HTML page by sending X-Requested-With header.
The result of this call is placed into a container, which is then shown to the user.

test_user_admin was updated to reflect updates in views.