HUE-799 [ua] Fix delete group action.
The list_groups.mako template had a javascript function that created a popup.
The template for the popup was taken and put into delete_group.mako.
Then, the delete_group view was updated to include appropriate context vars.
New javascript code was added to appropriately call an HTML page by sending X-Requested-With header.
The result of this call is placed into a container, which is then shown to the user.

No new tests were added. We need to consider more javascript oriented tests.