[beeswax] Integration of Hive Server 2
Add a layer of abstraction called 'dbms'. This layer is then calling
either a 'Beeswaxd' or 'Hive Server 2' service.

Beeswax is working as before. Hive Server 2 can now:
- list tables
- describe the columns of a table and provide and example of data
- send a SELECT query

Other features are not implemented yet. This is not part 1.

Hive Server 2 can be selected with the
server_interface=hiveserver2 property in [beeswax]