HUE-878 [desktop] Add a remote user backend
This backend enable to better support running Hue behind a proxy server.

Created a simple middleware to fix a bug in Django's built-in remote user
middleware that properly sets the header name to have the HTTP_ prefix and
added a backend that authenticates or creates a user given nothing but the
header. I also had to modify the settings.py to load the new middleware class.

Adding sample configuration to the hue.ini and pseudo-distributed.ini.tmpl files
Adding a comment to describe the purpose of the new middleware
Updating the documentation on the remote_user_header configuration setting