[fb] Upload to federated cluster
The location for the temporary file being uploaded should be
the destination of the final, renamed file. The temporary file
will be named <filename>.tmp.

No new tests were necessary since we are not changing upload
functionality. Simply changing where the temp files exist.

HACK: In order to get the destination when uploading in the
custom upload handler, needed to use a GET parameter. This
is due to the obscurity in Django's builtin POST data
parser. Basically, the 'dest' POST parameter is not accessible
until the file contents has been parsed.