HUE-967 [useradmin] Useradmin shouldn't depend on the existence of other apps
With transactional DBs, the useradmin app had to be the last app to sync.
With introspection, this is no longer the case. The useradmin app can be sync'd
at any point and start populating the HuePermission model.
The HuePermission model accepts strings, which makes it independent
of the model itself.