HUE-971 [oozie] Fix empty workflow
Start node should be able to point to end node if we replace its child with the end node.
The only time the start node will ever point to the end node is in the case of a brand new
workflow or if all nodes were removed. If all nodes are removed, the replaceChild
method is called. If we're start a workflow, we don't need to update the links of the start node.