HUE-697 [jobbrowser] Ability to browse MR2 jobs
Jobbrowser can now lists jobs and their details by using the YARN apis.
Currently, the Resource Manager API is used for listing the jobs and then
the MapReduce API is used when the job is running or otherwise the History
Server API is used. The current convention is to use application_ID when
the job is running and then job_ID when it is finished.

Compatible and tested with MR1 and MR2.
Tests added with YARN API mocked.
3 YARN API implemented and available in libs.
Retrieves logs with a curl as YARN API does not support it.
Asynchronously fetch the logs as it is slow.
With YARN, support a bew job_attempt page.
yarn_models.py created.
YARN API are all read only.
job URL updated in jobsub, oozie, beeswax.
Beeswax now displaying logs and the MR jobs ids.
Rename 'jobid' to 'job' in all the urls and make the check_job_permission
return the job object.
Added 3 new URL Hadoop YARN ini properties