HUE-933 [oozie] Fix clone node and unsaved ribbon
Unsaved ribbon was not going away because we were not setting the timeout
effectively. Using setInterval instead.

Clone node was not handling ribbon and state like edit and new node are.
Clone node follows the same logic as new node and edit node now.
First, the node is created and registered, but not added to the workflow.
If the node is successfully added (validated), then we add it to the workflow.
Otherwise, erase it if failed to validate and user cancels the clone process.