Fixing find_user when find_by_dn is True.
Setting distinguishedName= in the search filter is not a valid way of
searching by filter for an attribute.  distinguishedName is not a
valid attibute.  Instead, if we know the full dn of the user
we are searching for, just set search_dn to that it and don't
filter using user_name_attr.