[oozie] Import workflow with Subworkflow action and rudimentary support for all oozie schema version
Added a test case validating Subworkflow import.
In order to resolve foreign key dependency on workflow,
searching for deployment directory in HDFS, then iterating over
all workflows in DB. If one is found, the import will succeed.
Other wise, import will fail.
Made all oozie schema versions available for import.
The only major difference is in 0.1 where SSH action is not
an extension. It is a built in node.