[oozie] Drag a node onto workflow editor.
When adding a new node, users are now able to drag the node onto the canvas.
In order to save the workflow, the user will likely have to edit the node first.
Catching the 'dragstop' event, which should be called after 'drop'.
Cannot handle the 'drop' event since the workflow editor must use them and stops event bubbling.