HUE-159 [beeswax] Support multiple statements per query
Queries can now be composed of multiple statements separated by some semicolon.
As in the Hive shell, '\059' can be use in case of conflicts.
When the statement does not create a result set (e.g. CREATE, LOAD, DROP, ALTER)
the watch query page will just move automatically to the next statement when
the current one is finished. Moving to the next one loses the log of the previous one.
If the query has a dataset like with a SELECT, the result page will be showned at the
end and a 'continue' button in a note will invite the user to continue the execution
of the following statements.
If a statement has an error, similarly to a statement with a dataset, the result page
is showned and the user can also skip.
3 tests added for checking multi-statement queries with datasets, no datasets or a mix.
Query box is higher now and with a note about this new functionality.
Now pressing 'enter' on execute button will execute the query.
statement_number is a new field are added in QueryHistory in order to persist which
statement is currently running.