HUE-1094 [fb] Fail early when creating files in a folder without permissions
Do the upload as the current user as we can authenticate it before and are uploading
to the destination path.
Upload a file gets triggered by the is_ajax middleware. request.is_ajax() is not set
for some reason.
Upload archive works as expected but could not spot the different. This could be useful
for bubbling up the exception in the notification.