[core] JT plugin HA
Try to find another available configured MR1 cluster with a RUNNING status
in case of failure to connect to the JT plugin.
A JT in standby mode does not have its plugin listening.
A JT in active mode has its plugin listening.
Add a list of MapReduce cluster to hue.ini for listing the other standby JT.
For JobBrowser and Oozie.