HUE-1240 [beeswax] Save a query result when using HiveServer2 interface
Save result as a table with HiveServer 2 interface
HUE-927 [beeswax] Save result from non MR query to HDFS
Backward compatible with Beeswax ctas
Will use hive INSERT OVERWRITE for saving into HDFS even when using Beeswax
Data written to the filesystem is serialized as text with columns separated
by ^A and rows separated by newlines. If any of the columns are not of
primitive type - then those columns are serialized to JSON format.