[sqoop2] Fix databind errors when saving and running first job and connection
When a user performs the following steps, the app would have databind errors:
1. Delete all connections and jobs, then refresh the page.
2. Create a new connection with an error, then rectify the error, then save the connection.
3. Create a new job with an error, then rectify the error, then save and run the job.

There is a moment when switching between the job that was being edited and the newly loaded
job that would cause those errors. Also, needed to make sure "isLoading" is set when fetching
new jobs and connections.