HUE-1521 [core] Improve JT HA
We need the username of the current user when trying to find a new JT HA.
Checking if MR1 or MR2 should not trigger a call for checking the JT HA,
it should just look in the hue.ini cluster config.
When logged-in as a hue admin, the page might need to be refreshed once
because of the /debug/check_config_ajax call.