HUE-1545 [core] Improve SAML configuration parameters
acs, ls, and entity id are generated.
acs and ls urls can be generated because they are URLs in Hue.
entity id is generated because it's common to use
subdomain plus the url to the metadata view.
A hostname of 0.0.0.0 in hue.ini will result in
FQDN being used in acs, ls, and entity id.