HUE-1455 [impala] Close use and metastore queries
All operations are closed automatically except query() which is closed
by javascript when the user leaves the page.

We fetch 1000 rows for the other operations. We could have built
smart dataset that fetch/generate the data and then offer a close()
method but for now it should be fine.

Some operations that are watched (like drop tables, select from a partition,
create table as select, insert overwrite...) are closed like regular
query statements

Conflicts:

	apps/beeswax/src/beeswax/server/dbms.py
	apps/beeswax/src/beeswax/server/hive_server2_lib.py
	apps/beeswax/src/beeswax/views.py