HUE-1915 [core] MultipleObjectsReturned for tags or permissions
Only a strict unique clause in a DB could prevent duplicate tags or permissions
to be created. We clean them up at access time instead of filing as it is safe
to delete the duplicates.

Strick DB that fails when installing the Hue example should click on each individual
example install insteaf of the install all button.