HUE-1937 [beeswax] Clarify behavior of a multi query with error
When a multi query as an error, the user can modify the query and
click next to try to re-execute it without re-executing the queries
before.
The frontend was changed to transmit the full query each time between
two next statement. We also don't increase the statement number in
case of failure before.
We could also update the query between two statements but are not
doing it.

Fix in Oozie:
A PREP workflow should have the kill option
A coordinator workflow can become non in rare cases