[core] Fix dump_config tests by testing an app setting instead of desktop
With the loading of the desktop config first done when adding the
app blacklist, the set_for_testing() seems to be lost when dumping
the global setting variable.
This is only on the dump_config page which uses the global variable.
set_for_testing() continues to work fine in apps, desktop or libs.

Just indentation cleaning on the dump_config template.