HUE-1811 [search] Sample collections
Create 3 examples of collections with several facets and features:
- twitter
- yelp
- log search

The index page of the collections works even without a valid solr for demo purpose.
The collections are moved from the search dropdown to the main menu bar.
Some sample data is available for filling a reald index. Some basic indexing scripts
are in progress in order to offer a good quick start experience.

We are using hardoced ids in the fixtures to keep it simple.

Todo:
- click on install examples installs the example
- add more data and finish the indexing scripts
- collection level permissions
- alert message when using a demo collection without a valid solr