[pig] Logs can be missing with Yarn
So one of the pig problem is that because yarn logs disappear before re-appearing in History Server after a few s and we poll every 1s

Bumping the refresh time to 0.5 instead of 1s seems to solve the problem until Yarn improves or we put a smarter solution.

I tried to keep pooling until 10s in order to get the last part of the logs from the HS but then it bugs if the user runs the script in the meantime!
    $(document).on("stopLogsRefresh", function () {
      window.setTimeout(function () {
        window.clearInterval(logsRefreshInterval);
        $.jHueTitleUpdater.reset();
      }, 10000);
    });