HUE-1298 [impala] Multiple Impalad support
Made the DBMS.get method a factory that reuses instances on a per user per interface basis.
Impala sessions close when connectivity is lost.
Load balancing multiple impalad's with a persistent connection
enables Hue to communicate with the same impalad for as long as Hue is running
and the connection is alive.