HUE-2039 [beeswax] Remove the need to clone a shared query
A non shared design can't ne seen or modified by someone else without the
write perm.
A shared design can be executed in auto mode and will appear in the history
of the user.
A shared design can be saved as a new design.
Adding tests, refactored a bit the view to remove some logic from the view.