[core] Resource Manager RM HA
The [[yarn_clusters]] cluster configs in Hue ini need to be filled up with
at least 2 configs to configure Hue for RM HA.
submit_to must be True to have the RM taken into consideration.

The initial RM being used is the one named [[[default]]].
When submitting to Oozie, RM configs must specify the same logical_name.
When browsing JobBrowser, Job Browser will switch to the active RM on the fly if needed.