HUE-2446 [oozie] Add a no-op migration script to help upgrade from CDH4.7 to 5.x
A migration was backported from CDH 5 to CDH 4.7, which causes issues
when that instance of Hue is then upgraded to CDH 5. This puts a
bandaid on the problem by adding a no-op migration script that silences
the south migration.

In the future, it would be better if we add support for adding explicit
migrations from version X of Hue to verion Y, so that we can clean up
databases that have gotten into an odd state.