HUE-2485 [core] Add strict redirection in EnsureSafeRedirectURLMiddleware
Hue has a security vulnerability that redirectiond to an arbitrary site
is allowed after a successful login if the '?next=' variable is modified
by a malicious hacker. This vulnerability can be used to mount a phishing
attack against the cluster where Hue is running.

This fix sets redirect_whitelist configuration default value to "^\/.*$",
and limits all redirections to the same host where HUE is running by default
to prevent phishing attacks.