HUE-2436 [oozie] Django-1.6: Change massaged_oozie_jobs_for_json to return dict
Django-1.6 now raises an exception if you create a json response
and the `safe=True` flag is not set. This is because on some older
browsers it's possible to replace the JavaScript Array constructor
to do bad things:

http://django.readthedocs.org/en/latest/ref/request-response.html#serializing-non-dictionary-objects

One odd thing to note is we don't appear to be using:

list_oozie_coordinators?format=json
list_oozie_bundles?format=json

so we could probably delete this from the functions.