HUE-2436 [oozie and pig] Don't cache the oozie api, this can break the tests
Prior to this patch, the `get_oozie` cached a connection to oozie.
However, the tests spin up their own oozie with a different url than
the default oozie, so there is a race on who creates the connection.

It turns out the `OozieApi` is pretty cheap to create, so instead of
a global cache, this PR changes the requests to create one `oozie_api`
and thread it through to all the users of that connection.