HUE-2436 [desktop] Protect tests from falling into an infinite loop
The mako test runner doesn't handle one template calling another.
This can happen if when rendering one template we call a function
that throws an exception, triggering the rendering of the error
page. This results in an infinite loop and obscures the original
failure. This patch instead builds a stack of template render calls
so that the original error can be discovered.