[desktop] Fix migrations for postgresql
This migration referenced the "Document" table before it was created.
It also called Document.object.sync() before those tables may have
been created. So instead of failing the whole transaction (which then
fails the migration), it'll just put syncing the other models in their
own sub-transaction and continue on if they fail.