HUE-1660 [core] API for export/import all stored scripts
Export or import Hue documents2. Implemented genericly and for Spark Notebooks.

If a document uuid already exists it will be replaced. A non admin cannot export documents he does not own (later will extend to document he can read),
and cannot import documents of someone else.

Exports are done by using the id of the documents, this keeps the data simple.
Imports are done by using the uuid of the documents, this guarantees that each document is unique, independently of its DB id.

Export
------
/desktop/api2/doc/export?documents=[50093,50091]

Parameters:
Json list of pk of documents2
Returns:
json file (default)
?format=json: json response
?format=zip: zip file with json data and text representation of compatible documents, e.g. oozie (later), SQL (later), Notebooks...

Import
------
/desktop/api2/doc/import
POST:
With a json document named 'documents', same as the one coming from export in json.

or

FILE:
With a file named 'documents' that containts the content of some exported documents in json.
With a 'redirect' input.

Examples
-------
Examples in notebooks.mako