[libzookeeper] Move the ZooKeeper specific properties to its own lib
The goal is to move ZooKeeper infos into a lib that is always available even if
the ZooKeeper app is disabled.

Indexer, Sentry needs to know some ZooKeeper properties.

The ZooKeeper app will use libzookeeper properties in Hue 4. Only its
app specific properties should be there.

This is similar to liboozie which is used by the Oozie, JobSub and Pig apps.

Conflicts:

	desktop/Makefile
	desktop/libs/indexer/src/indexer/conf.py