[notebook] Add code visibility toggle to snippets
This change makes it possible to show and hide the code or properties area for a snippet. Before this change the run and stop buttons were placed on top of the code editor and in order to still allow execution when the code is hidden the button has been moved down below the editor.

The visibility state is saved with the notebook.