[spark] Move assist panel to a separate KO component
With this commit the assist panel code is moved to a separate KO component. A new ko_components.mako is created to act as a library for this and future knockout components.

The assist instance is currently shared with codemirror for autocomplete so it currently has to be created from outside the component and passed as a reference. When we move to a new editor we should consider refactoring this as well.