[desktop] Fix hue suppressing psycopg2 import error
This fixes a problem where hue reports
`Unknown command: 'runcherrypyserver'` when it's configured to talk
to a postgresql database, but psycopg2 isn't installed inside hue.
It turns out that importing `desktop.monkey_patches` in settings
will result in django eventually trying to import the database library,
which raises an `ImproperlyConfigured` error. However, django
ignores `ImproperlyConfigured` when it's building it's list of supported
user commands, so we never would see the import error.

This patch fixes the immediate problem, but it doesn't yet stop this
from happening again in the future.