HUE-2851 [build] Pickup system packages
Due to licensing issues, Hue is unable to bundle in certain dependencies,
like psycopg2. This patch allows Hue admins to install packages in
the global python site-packages that will be exposed to Hue. The bundled
Hue dependencies take precedence over the global dependencies to make
sure we are not using an unexpected version.

This does expose Hue to some risk:

1. Hue developers could accidentally introduce a dependency on a
package local to their system. The unit tests should detect and
protect against it.

2. Since we have no control over what packages the end user installs
for themselves, they could install a version that is incompatible
with Hue. Errors like this should be caught by doing matrix testing
with all the Hue supported platforms.