[spark] Improved css flow for the notebook panels
This commit changes the css structure for the notebook contents, moving away from the use of fixed positioning while still allowing the assist panel to "stick". By doing this it will be easier to add resize functionality to the panels.

The only visual change is that scrolling now takes place inside the div containing the snippets, before it was for the entire page.