[core] Switch to pixel widths for ko splitDraggable
This removes the relative positioning and sizing of the panels and instead uses fixed pixels for the active positioning. Before this the panels used a mix of percent and pixels for the layout model which caused undesired effects for extra narrow and extra wide browser windows. By switching to pixels we can easier control the presentation.

This also moves some of the resize logic to inside the binding, getting rid of the need for an external observable that keeps track of the left panel width.