[spark] Move left panel toggle out of assist component
As the layout is controlled from the page it doesn't make sense that the assist component should be involved. For the notebook this has been fine as the assist is the only contents of the left panel but for apps like beeswax there's other stuff there as well, so this change will also make it easier to replace the beeswax assist with this component later.

The icon is now placed in the upper right corner of the left panel instead of next to the database header as before.