HUE-2727 [spark] Fix session restart
This adds a dedicated function for restarting the sessions. Instead of removing/adding a new one it keeps the session and updates it when the new one is created server-side.

When a notebook is loaded with existing sessions it will now restart those session on first execute instead of creating a completely new session as before.