HUE-2864 [livy] Initial support for sparkr interactive sessions
This adds basic support for running sparkr interactive sessions.
It does this by shelling out to "sparkr" and parsing the output.
This has some limitations. First of all, in YARN mode it is not
yet able to reuse the application master like the scala and
python sessions, so it allocates a redundant master. Second,
it does not yet acknowledge the spark configuration settings.