[spark] Fix snippet type switch
When the snippet specific properties were introduced it broke the snippet type switch. With this commit it will update the properties of the snippet to match the type, it will also keep any previous properties for the snippet type in case the user changes his mind and switches back to the initial type.