[spark] Clone the initial values in the ko csv input component
If the csv input component is called with an array of observables the observables will be updated with the internal structure of the csv component, which is not what we want. The error appears after loading a saved notebook containing snippets with some values set for the csv properties, it's caused by the ko mapping plugin creating observables for these values.

This fixes the issue by using the ko mapping toJS function to ensure that no observables are referenced.