HUE-2664 [jobbrowser] Fix fetching logs from job history server
While the job is active we need to hack up the url we're receiving
from YARN in order to point at the right location. However this
is wrong to do after the job has been migrated to the history server.
So this patch just skips modifying the url if it happens to be
pointed at the job history server.