HUE-2908 [livy] Partially address multiline commands
This allows for a R snippet to be spread across multiple lines, but
it requires that the entire R snippet is a whole statement, or
multiple statements. If a statement is spread across multiple snippets,
then the repl will hang. Unfortunately the only real way to handle
this is to parse the R command for correctness, or just do the
proper thing and write a fake_shell.py-ish thing for R.