HUE-2852 [core] Support autocomplete of struct types
This adds support for struct autocompletion between SELECT and FROM as well as after WHERE, ON and ORDER BY

The completion after WHERE, ON and ORDER BY was also broken for normal column completion and this commit fixes that as well.

I've also changed the caching to one cache entry for the autocomplete where the urls are used as keys in the cached data. This allows us to invalidate the entire cache in one go.