[core] Support autocomplete for array types in hive mode
This adds autocomplete support for array items. The api url to fetch the terms is slightly different than for maps so we have to make additional calls to fetch the type before we can give the suggestions. The calls go through the assistHelper so any previous responses for the same url are cached and returned synchronously.