[libzookeeper/indexer] Rewrite to not leak sockets or temp files
We weren't closing the zookeeper in a couple places, so to be
safer, this patch changes all users of ZookeeperClient to use
a `with` context to make sure things get closed.

Along the way, it adds some try-excepts and try-finallys to make
sure all the temp directories created by `copy_configs` are
always cleaned up.