[spark] Fix cache issue with assistHelper
If the database part of the API URL is undefined or null any calls will still come back as a a success, a code is instead used which could represent an error. This commit will prevent caching of a response with code 500. I've also added an additional check to make sure there's an active database before fetching the tables in the ace binding.
(cherry picked from commit 82a3ddf16110da8ad589ef9189b6376d4832ddd5)