[core] Improved handling of click and double-click on the same element
When combining click and double-click events on the same element (or element hierarchy) the double click will also trigger a click. To prevent this I've created a knockout binding that will only trigger the double click (if quick enough).

Combining click and double click events is not very good UX so this should be seen as a temporary solution until we find a better way.