[core] Fix incorrect table alias suggestion in autocomplete
This fixes an issue where parts of a nested select appear in the table alias suggestions. With this fix it will ignore any identifiers that contain a '('.

It's definitely possible to suggest aliases of nested select statements but it's quite complex with the current regex-based parser, so for now I've only added a disabled test for it.