[notebook] Improved multi-snippet autocomplete
This takes care of an issue where all autocompleters are called for all snippets when editing a single snippet. With the Ace support for session autocompleters it will now only call the completer that belongs to the snippet.

This also allows us to drop some of the odd work-arounds, and I'v dropped the extraCompleters from the binding as it doesn't seem to be in use anymore.