[impala] Check if creating the impala test databases had failed
On Kerberos, it's possible for the tests to fail with some
mysterious errors just say that a QueryHistory object had
not been created. This is happening because the Hue test
user may not necessarily have permission to write to these
tables. This patch doesn't fix the error, it just surfaces
the real problem being that creating these tables is failing.