[hue] Fix parsing classes in pyspark
This changes how python code is compiled. Before, we compiled
line by line until a chunk compiled. This was done in order to
catch magic lines while still allowing for '%' to be used inside
a multiline string as in:

```
x = """
%string"""
%json x
```

This would technique would properly return the json value of
`{"text/plain": "\n%string"}`.

Unfortunately this approach doesn't work with classes with
multiple methods, because a class with a single method
is syntatically valid. This patch fixes this bug by parsing
the whole snippet at once, and only if there's a syntax error
will it try to see if there's a magic command in the
snippet.