[core] Fix pre-selection issue with the select2 binding
This takes care of an issue that occur when the value and options of a select2 binding are independently updated in the same ko update cycle. If the selected value is updated before the options the select2 binding changes the selected value to empty, overwriting the new value, and once the options are updated the original value is no longer there.

With this change it will now refresh both the options and the selected value at the same time when any of them mutate.