[oozie] Load the ko.hue-bindings from the main pages
The html of the submit modal is generated dynamically and the contents is rendered using jquery html(). For the submit modal it contains additional <script> resources and I believe this is not always picked up in time (or at all). To ensure it's there I've now added the ko.hue-bindings where knockout is loaded.