Lining up Spar/Livy app maven versions.
Maven (i.e., "mvn install" in apps/spark/java) was having trouble building,
because the versions didn't line up.  With this change, the verions are
3.9.0-cdh5.7.6-SNAPSHOT all across the board.  I chose this (rather than
0.2.0-cdh5.7.6-SNAPSHOT) because
http://maven.jenkins.cloudera.com:8081/artifactory/cdh-snapshot-local/com/cloudera/hue/livy/livy-main/
suggested that this has existed for past versions.

I'm a little bit surprised that this was building in CDH-land at all, but it
looks like the non-cdh-specific version was populated in repos like
http://maven.jenkins.cloudera.com:8081/artifactory/cdh-snapshot-local/com/cloudera/hue/livy/livy-main/.