[assist] Make the assistHelper singleton to fix caching issues
There's some state kept in the assist helper to simplify the autocompleter, this change guarantees that also this gets cleared when the assist is refreshed.

This also removes the index of assist helpers by type from the notebook (the type doesn't matter anymore to the assist helper), by doing so it also fixes another caching issue with the notebook snippets that use the helper.