[doc2] Import Hive and Impala queries from document1 to document2
Old format queries are not deleted, similarly to Oozie v1 to v2.
The import happens by user when then go on the new home page.
When a document 1 is imported, it will be tagged so that it is not re-imported again.
Pig and Job Designers docs will be linked shortly with a doc2.
History docs are not imported for now.
Tests are coming up when it stabilizes.