[assist] Improved cache refresh handling
With this change it will no longer clear the complete cache on refresh but instead just for the active source type (hive, impala etc.), the change also paves way for HDFS cache refresh.

This will also fix a couple of issues with the metastore. The cache was cleared twice when dropping tables and DBs causing multiple reloads and there was inconsistent types of icons in the create table wizards.